requirements.txt
Appium-Python-Client==0.26
astroid==1.6.3
attrs==17.4.0
bert==2.2.0
certifi==2018.4.16
chardet==3.0.4
cx-Oracle==6.2.1
DAWG-Python==0.7.2
decorator==4.3.0
docopt==0.6.2
erlastic==2.0.0
Faker==0.7.7
fire==0.1.3
humanize==0.5.1
idna==2.6
isort==4.3.4
lazy-object-proxy==1.3.1
mccabe==0.6.1
more-itertools==4.1.0
paho-mqtt==1.3.1
pluggy==0.6.0
progress==1.3
py==1.5.3
pylint==1.8.4
pymorphy2==0.8
pymorphy2-dicts==2.4.393442.3710985
pytest==3.5.0
python-dateutil==2.7.2
python-dotenv==0.8.2
requests==2.18.4
retry==0.9.2
selenium==3.11.0
six==1.11.0
text-unidecode==1.2
tornado==5.0.1
uiautomator2==0.1.0
urllib3==1.22
wrapt==1.10.11
